Refactor DonateFiatThanks into a reusable component
